<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

Forced installation of packages, ignoring peer dependency warnings from npm, because we're not installing peer dependencies anyway.

### Reason

`react-native` doesn't yet support React 19

https://github.com/facebook/react-native/blob/main/packages/react-native/package.json#L102

However that is what our dependencies install because we have a peer on `^18.0.0 || ^19.0.0-beta`.
And package benching uses `npm`, which is strict on peer deps mismatch.

![image](https://github.com/user-attachments/assets/1ff70a4c-43c4-40cf-bf1e-0aa927462d01)


This is usually not a problem, because projects have explicit dependencies on React 18:

```json
"react": "18.3.1",
"react-dom": "18.3.1"
````

That works, because all packages have an overlap with that. Users will always have these dependencies, and so does the sandboxes. But our virtual bench packages _don't_ depend on `react`, so npm chooses the highest of `^18.0.0 || ^19.0.0-beta` which is `19.0.0`.

So another solution would be to add `overrides` with React 18, but just using force seems more future proof, we'd 100 % forget to change this when upgrading the monorepo to React 19.

<!-- Briefly describe what your PR does -->

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  77.7 MB | 77.7 MB | -4.28 kB | **-3.73** | 0% |
| initSize |  130 MB | 130 MB | 14.7 kB | -1.2 | 0% |
| diffSize |  52.4 MB | 52.5 MB | 19 kB | **1.94** | 0% |
| buildSize |  6.75 MB | 6.75 MB | 0 B | -0.65 | 0% |
| buildSbAddonsSize |  1.51 MB | 1.51 MB | 0 B | - | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.86 MB | 1.86 MB | 0 B | **1.33** | 0% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | -0.65 | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.57 MB | 3.57 MB | 0 B | -0.65 | 0% |
| buildPreviewSize |  3.19 MB | 3.19 MB | 0 B | 0.65 | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  6.8s | 7.9s | 1.1s | -0.9 | 14.4% |
| generateTime |  25.5s | 23.2s | -2s -250ms | **1.56** | 🔰-9.7% |
| initTime |  15.9s | 17.7s | 1.8s | **2.31** | 🔺10.4% |
| buildTime |  8.1s | 9.5s | 1.4s | 0.41 | 14.8% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5.4s | 5.2s | -136ms | 0.13 | -2.6% |
| devManagerResponsive |  3.8s | 3.8s | -1ms | 0.31 | 0% |
| devManagerHeaderVisible |  555ms | 630ms | 75ms | 0.7 | 11.9% |
| devManagerIndexVisible |  629ms | 664ms | 35ms | 0.17 | 5.3% |
| devStoryVisibleUncached |  1.9s | 1.8s | -36ms | 0.55 | -1.9% |
| devStoryVisible |  627ms | 666ms | 39ms | 0.24 | 5.9% |
| devAutodocsVisible |  408ms | 573ms | 165ms | **1.24** | 🔺28.8% |
| devMDXVisible |  562ms | 525ms | -37ms | 0.04 | -7% |
| buildManagerHeaderVisible |  549ms | 527ms | -22ms | -0.49 | -4.2% |
| buildManagerIndexVisible |  622ms | 604ms | -18ms | -0.26 | -3% |
| buildStoryVisible |  498ms | 489ms | -9ms | -0.56 | -1.8% |
| buildAutodocsVisible |  440ms | 470ms | 30ms | 0.32 | 6.4% |
| buildMDXVisible |  404ms | 426ms | 22ms | -0.2 | 5.2% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Based on the provided information, I'll create a concise summary of the PR changes:

Modified package benchmarking script to handle React version compatibility issues by forcing npm installations to ignore peer dependency warnings.

- Added `--force` flag to npm install command in `scripts/bench/bench-packages.ts` to bypass peer dependency checks
- Addresses react-native compatibility issue with React 19 beta versions
- Avoids need for explicit React 18 overrides in benchmark packages
- Future-proofs solution for eventual React 19 upgrade
- Maintains existing benchmark functionality while fixing @storybook/react-native-web failures



<!-- /greptile_comment -->